requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.2.1
python-dotenv==1.0.0
beautifulsoup4==4.12.3
requests==2.31.0